LayoutTests/svg/animations/svgtransform-animation-discrete-expected.txt

SVG 1.1 dynamic animation tests

Test calcMode=discrete animation on SVGAnimateTransform.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.transform.animVal.numberOfItems is 1
PASS rect.transform.animVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_TRANSLATE
PASS rect.transform.animVal.getItem(0).matrix.e is 100
PASS rect.transform.animVal.getItem(0).matrix.f is 100
PASS rect.transform.baseVal.numberOfItems is 0
PASS rect.transform.animVal.numberOfItems is 1
PASS rect.transform.animVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_TRANSLATE
PASS rect.transform.animVal.getItem(0).matrix.e is 100
PASS rect.transform.animVal.getItem(0).matrix.f is 100
PASS rect.transform.baseVal.numberOfItems is 0
PASS rect.transform.animVal.numberOfItems is 1
PASS rect.transform.animVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_TRANSLATE
PASS rect.transform.animVal.getItem(0).matrix.e is 0
PASS rect.transform.animVal.getItem(0).matrix.f is 0
PASS rect.transform.baseVal.numberOfItems is 0
PASS rect.transform.animVal.numberOfItems is 1
PASS rect.transform.animVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_TRANSLATE
PASS rect.transform.animVal.getItem(0).matrix.e is 0
PASS rect.transform.animVal.getItem(0).matrix.f is 0
PASS rect.transform.baseVal.numberOfItems is 0
PASS rect.transform.animVal.numberOfItems is 0
PASS rect.transform.baseVal.numberOfItems is 0
PASS successfullyParsed is true

TEST COMPLETE
